requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-countries==7.4.2
django-js-asset==2.0.0
django-mptt==0.14.0
Pillow==9.3.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
